Removed most of the obsolete messages flushing the text screen. Worked a bit to improve the mission files.

+++ src/struct.h
@@ -145,7 +145,7 @@ influence_t, *Influence_t;
 
 typedef struct
 {
-  byte type;			/* gibt die Nummer in Druidmap an */
+  int type;			/* gibt die Nummer in Druidmap an */
   int levelnum;			/* Level in dem sich enemy befindet */
   finepoint pos;		/* gibt die Koordinaten der Momentanposition an */
   finepoint speed;		/* current speed  */
